biosight/api/requirements.txt
google-generativeai
python-dotenv
pillow
fastapi
uvicorn
tenacity
python-multipart
